Multiple call function added
